리뷰 반영: Players 에서 파싱 로직 분리
